Playerが死亡するとGameOverに遷移するように実装
※PlayerがBossを倒した後にPlayerのHPが0になるとGameOverになってしまう不具合を修正

+++ source/UI/PlayerHUD.hpp
@@ -53,7 +53,7 @@ private:
 	int buff_amount;
 
 	float player_hp;
-	float player_state;
+	int player_state;
 	Vec2 player_vel;
 	float player_avoid;
